split out perf-now
